#12 Refactor errors.Make function into its own file
